+++ browse/src/sse-session-cookie.ts
@@ -21,29 +21,22 @@
  * - In-memory only. No persistence across daemon restarts — extension
  *   re-mints on reconnect.
  * - Tokens are 32 random bytes (URL-safe base64). 256 bits, unbruteforceable.
+ *
+ * Shares the registry implementation with pty-session-cookie.ts via
+ * createSessionCookieStore; separate INSTANCE so the token spaces never
+ * overlap.
  */
-import * as crypto from 'crypto';
-
-interface Session {
-  createdAt: number;
-  expiresAt: number;
-}
+import { createSessionCookieStore } from './session-cookie-store';
 
 const TTL_MS = 30 * 60 * 1000; // 30 minutes
-const MAX_SESSIONS = 10_000; // Upper bound on registry size
-const sessions = new Map<string, Session>();
 
 export const SSE_COOKIE_NAME = 'gstack_sse';
 
+const store = createSessionCookieStore({ cookieName: SSE_COOKIE_NAME, ttlMs: TTL_MS });
+
 /** Mint a fresh view-only SSE session token. */
 export function mintSseSessionToken(): { token: string; expiresAt: number } {
-  // 32 random bytes → 43-char URL-safe base64 (no padding)
-  const token = crypto.randomBytes(32).toString('base64url');
-  const now = Date.now();
-  const expiresAt = now + TTL_MS;
-  sessions.set(token, { createdAt: now, expiresAt });
-  pruneExpired(now);
-  return { token, expiresAt };
+  return store.mint();
 }
 
 /**
@@ -53,68 +46,20 @@ export function mintSseSessionToken(): { token: string; expiresAt: number } {
  * unboundedly under sustained mint + reconnect pressure.
  */
 export function validateSseSessionToken(token: string | null | undefined): boolean {
-  if (!token) return false;
-  const s = sessions.get(token);
-  if (!s) {
-    pruneExpired(Date.now());
-    return false;
-  }
-  if (Date.now() > s.expiresAt) {
-    sessions.delete(token);
-    pruneExpired(Date.now());
-    return false;
-  }
-  return true;
+  return store.validate(token);
 }
 
 /** Parse the SSE session token from a Cookie header. */
 export function extractSseCookie(req: Request): string | null {
-  const cookieHeader = req.headers.get('cookie');
-  if (!cookieHeader) return null;
-  for (const part of cookieHeader.split(';')) {
-    const [name, ...valueParts] = part.trim().split('=');
-    if (name === SSE_COOKIE_NAME) {
-      return valueParts.join('=') || null;
-    }
-  }
-  return null;
+  return store.extract(req);
 }
 
-/**
- * Build the Set-Cookie header value for the SSE session cookie.
- * - HttpOnly: not readable from JS (mitigates XSS token exfiltration)
- * - SameSite=Strict: not sent on cross-site requests (mitigates CSRF)
- * - Path=/: scope to the whole origin so SSE endpoints can read it
- * - Max-Age matches the TTL
- *
- * Secure is intentionally omitted: the daemon binds to 127.0.0.1 over
- * plain HTTP, and setting Secure would prevent the browser from ever
- * sending the cookie back. If gstack ever ships over HTTPS, add Secure.
- */
+/** Build the Set-Cookie header value for the SSE session cookie. */
 export function buildSseSetCookie(token: string): string {
-  const maxAge = Math.floor(TTL_MS / 1000);
-  return `${SSE_COOKIE_NAME}=${token}; HttpOnly; SameSite=Strict; Path=/; Max-Age=${maxAge}`;
-}
-
-function pruneExpired(now: number): void {
-  // Opportunistic cleanup: check up to 20 entries per call so we don't
-  // stall on a massive registry. O(1) amortized.  Runs on every mint
-  // AND on every validate so a steady reconnect flow can't outpace it.
-  let checked = 0;
-  for (const [token, session] of sessions) {
-    if (checked++ >= 20) break;
-    if (session.expiresAt <= now) sessions.delete(token);
-  }
-  // Hard cap as a backstop — if something still gets past opportunistic
-  // cleanup (e.g., all unexpired but registry enormous), drop the oldest.
-  while (sessions.size > MAX_SESSIONS) {
-    const first = sessions.keys().next().value;
-    if (!first) break;
-    sessions.delete(first);
-  }
+  return store.buildSetCookie(token);
 }
 
 // Test-only reset.
 export function __resetSseSessions(): void {
-  sessions.clear();
+  store.__reset();
 }
